LayoutTests/canvas/philip/tests/2d.state.saverestore.shadowBlur-expected.txt

< [index] >
2d.state.saverestore.shadowBlur
save()/restore() works for shadowBlur
References: 2d.state.shadowBlur
Actual output:
Passed
